requirements.txt
click
imageio
pillow
